chore: split run into multiple files
